Qt: add shortcut manager

+++ rpcs3/Emu/NP/rpcn_config.cpp
@@ -65,7 +65,7 @@ std::string cfg_rpcn::generate_npid()
 	const char list_chars[] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
 		'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'};
 
-	std::srand(time(nullptr));
+	std::srand(static_cast<u32>(time(nullptr)));
 
 	for (int i = 0; i < 10; i++)
 	{
